relief_engine/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
numpy==2.0.1
requests==2.32.3
scipy==1.14.0
trimesh==4.4.9
